+++ rstsr-common/src/lib.rs
@@ -6,13 +6,13 @@
 pub mod prelude;
 pub mod prelude_dev;
 
+pub mod alloc_vec;
 pub mod axis_index;
 pub mod error;
 pub mod flags;
 pub mod format_layout;
 pub mod layout;
 pub mod util;
-pub mod alloc_vec;
 
 #[cfg(feature = "rayon")]
 pub mod par_iter;
