+++ rstsr-core/src/feature_rayon/auto_impl/op_tri.rs
@@ -1,4 +1,3 @@
-use crate::feature_rayon::*;
 use crate::prelude_dev::*;
 use num::complex::ComplexFloat;
