--- rstsr-core/src/feature_rayon/operators/mod.rs
@@ -1,5 +1,0 @@
-pub mod op_tri;
-pub mod op_with_func;
-
-pub use op_tri::*;
-pub use op_with_func::*;
