+++ rstsr-native-impl/src/cpu_rayon/op_tri.rs
