+++ rstsr-native-impl/src/cpu_rayon/op_with_func.rs
