+++ rstsr-native-impl/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod prelude_dev;
 
+pub mod cpu_rayon;
 pub mod cpu_serial;
